+++ wsgidav/wsgidav_app.py
@@ -3,7 +3,7 @@
 # Original PyFileServer (c) 2005 Ho Chun Wei.
 # Licensed under the MIT license:
 # http://www.opensource.org/licenses/mit-license.php
-"""
+r"""
 ::
 
      _      __         _ ___  ___ _   __
@@ -126,13 +126,11 @@ class WsgiDAVApp(object):
 
         self.config = copy.deepcopy(DEFAULT_CONFIG)
         util.deep_update(self.config, config)
-        # self.config.update(config)
         config = self.config
 
         # Evaluate configuration and set defaults
         _check_config(config)
 
-        #        response_trailer = config.get("response_trailer", "")
         self.verbose = config.get("verbose", 3)
 
         lock_storage = config.get("lock_manager")
@@ -154,12 +152,12 @@ class WsgiDAVApp(object):
         self.mount_path = config.get("mount_path")
         auth_conf = config.get("http_authenticator", {})
 
-        # Instantiate DAV resource provider objects for every share
-        # provider_mapping may contain the args that are passed to a `FilesystemProvider`
-        # instance:
+        # Instantiate DAV resource provider objects for every share.
+        # provider_mapping may contain the args that are passed to a
+        # `FilesystemProvider` instance:
         #     <mount_path>: <folder_path>
         # or
-        #     <mount_path>: { "root": <folder_path>, "readonly": False }
+        #     <mount_path>: { "root": <folder_path>, "readonly": True }
         # or contain a complete new instance:
         #     <mount_path>: <DAVProvider Instance>
 
@@ -170,6 +168,7 @@ class WsgiDAVApp(object):
         for share, provider in provider_mapping.items():
             self.add_provider(share, provider)
 
+        self.http_authenticator = None
         domain_controller = None
 
         # Define WSGI application stack
@@ -180,16 +179,16 @@ class WsgiDAVApp(object):
         # is eventually called by the server.
         self.application = self
 
-        # The `middleware_stack` is configured such that the first app in the list should be
-        # called first. But since every app wraps its predecessor, we iterate in reverse
-        # order:
+        # The `middleware_stack` is configured such that the first app in the
+        # list should be called first. Since every app wraps its predecessor, we
+        # iterate in reverse order:
         for mw in reversed(middleware_stack):
             # The middleware stack configuration may contain plain strings, dicts,
             # classes, or objects
             app = None
             if compat.is_basestring(mw):
-                # If a plain string is passed, try to import it, assuming BaseMiddleware
-                # signature
+                # If a plain string is passed, try to import it, assuming
+                # `BaseMiddleware` signature
                 app_class = dynamic_import_class(mw)
                 app = app_class(self, self.application, config)
             elif type(mw) is dict:
@@ -209,13 +208,10 @@ class WsgiDAVApp(object):
                 # Otherwise assume an initialized middleware instance
                 app = mw
 
-            # FIXME: We should generalize this specific code:
+            # Remember
             if isinstance(app, HTTPAuthenticator):
+                self.http_authenticator = app
                 domain_controller = app.get_domain_controller()
-                # Check anonymous access
-                for share, data in self.provider_map.items():
-                    if app.allow_anonymous_access(share):
-                        data["allow_anonymous"] = True
 
             # Add middleware to the stack
             if app:
@@ -224,6 +220,7 @@ class WsgiDAVApp(object):
             else:
                 _logger.error("Could not add middleware {}.".format(mw))
 
+        domain_controller
         # Print info
         _logger.info(
             "WsgiDAV/{} Python/{} {}".format(
@@ -252,23 +249,30 @@ class WsgiDAVApp(object):
         if self.verbose >= 3:
             _logger.info("Registered DAV providers by route:")
             for share in self.sorted_share_list:
-                data = self.provider_map[share]
-                hint = " (anonymous)" if data["allow_anonymous"] else ""
-                _logger.info("  - '{}': {}{}".format(share, data["provider"], hint))
+                provider = self.provider_map[share]
+                hint = (
+                    " (anonymous)"
+                    if domain_controller.is_share_anonymous(share)
+                    else ""
+                )
+                _logger.info("  - '{}': {}{}".format(share, provider, hint))
 
         if auth_conf.get("accept_basic") and not config.get("ssl_certificate"):
             _logger.warning(
                 "Basic authentication is enabled: It is highly recommended to enable SSL."
             )
 
-        for share, data in self.provider_map.items():
-            if data["allow_anonymous"]:
-                # TODO: we should only warn here, if --no-auth is not given
-                _logger.warning("Share '{}' will allow anonymous access.".format(share))
+        if domain_controller:
+            for share, provider in self.provider_map.items():
+                if domain_controller.is_share_anonymous(share):
+                    # TODO: we should only warn here, if --no-auth is not given
+                    _logger.warning(
+                        "Share '{}' will allow anonymous access.".format(share)
+                    )
         return
 
     def add_provider(self, share, provider, readonly=False):
-        """Add a provider to the provider_map."""
+        """Add a provider to the provider_map routing table."""
         # Make sure share starts with, or is '/'
         share = "/" + share.strip("/")
         assert share not in self.provider_map
@@ -311,7 +315,8 @@ class WsgiDAVApp(object):
         provider.set_lock_manager(self.lock_manager)
         provider.set_prop_manager(self.prop_manager)
 
-        self.provider_map[share] = {"provider": provider, "allow_anonymous": False}
+        self.provider_map[share] = provider
+        # self.provider_map[share] = {"provider": provider, "allow_anonymous": False}
 
         # Store the list of share paths, ordered by length, so route lookups
         # will return the most specific match
@@ -319,6 +324,29 @@ class WsgiDAVApp(object):
         self.sorted_share_list = sorted(self.sorted_share_list, key=len, reverse=True)
 
         return provider
+
+    def resolve_provider(self, path):
+        """Get the registered DAVProvider for a given path.
+
+        Returns:
+            tuple: (share, provider)
+        """
+        # Find DAV provider that matches the share
+        share = None
+        lower_path = path.lower()
+        for r in self.sorted_share_list:
+            # @@: Case sensitivity should be an option of some sort here;
+            # os.path.normpath might give the preferred case for a filename.
+            if r == "/":
+                share = r
+                break
+            elif lower_path == r or lower_path.startswith(r + "/"):
+                share = r
+                break
+
+        if share is None:
+            return None, None
+        return share, self.provider_map.get(share)
 
     def __call__(self, environ, start_response):
 
@@ -357,24 +385,27 @@ class WsgiDAVApp(object):
         environ["wsgidav.verbose"] = self.verbose
 
         # Find DAV provider that matches the share
-        share = None
-        lower_path = path.lower()
-        for r in self.sorted_share_list:
-            # @@: Case sensitivity should be an option of some sort here;
-            # os.path.normpath might give the preferred case for a filename.
-            if r == "/":
-                share = r
-                break
-            elif lower_path == r or lower_path.startswith(r + "/"):
-                share = r
-                break
+        share, provider = self.resolve_provider(path)
+        # share = None
+        # lower_path = path.lower()
+        # for r in self.sorted_share_list:
+        #     # @@: Case sensitivity should be an option of some sort here;
+        #     # os.path.normpath might give the preferred case for a filename.
+        #     if r == "/":
+        #         share = r
+        #         break
+        #     elif lower_path == r or lower_path.startswith(r + "/"):
+        #         share = r
+        #         break
 
         # Note: we call the next app, even if provider is None, because OPTIONS
         #       must still be handled.
         #       All other requests will result in '404 Not Found'
-        if share is not None:
-            share_data = self.provider_map.get(share)
-            environ["wsgidav.provider"] = share_data["provider"]
+        # if share is not None:
+        #     share_data = self.provider_map.get(share)
+        #     environ["wsgidav.provider"] = share_data["provider"]
+
+        environ["wsgidav.provider"] = provider
         # TODO: test with multi-level realms: 'aa/bb'
         # TODO: test security: url contains '..'
 
@@ -462,7 +493,7 @@ class WsgiDAVApp(object):
 
             # Log request
             if self.verbose >= 3:
-                userInfo = environ.get("http_authenticator.user_name")
+                userInfo = environ.get("wsgidav.auth.user_name")
                 if not userInfo:
                     userInfo = "(anonymous)"
                 extra = []